// src/app/pages/masters/patient/types.ts
export type PatientInsuranceType = {
    id: number;
    PatientID: number | null;
    PolicyNumber: string;
    PlanNumber: string;
    PlanType: string;
    EffectiveDate: string | null;
    ExpiryDate: string | null;
    Relationship: string;
    InsuranceProviderID: string;
    InsuranceProviderName: string;
    VerificationStatus: string;
    VerifiedStaffID: number | null;
    VerificationDate: string | null;
    VerificationNotes: string;
    InsurerNameAsInPolicy: string;
    Address: string;
    City: string;
    PostalCode: string;
    State: string;
    Country: string;
    RelationName: string;
    Relation_Address: string;
    Relation_City: string;
    Relation_PostalCode: string;
    Relation_State: string;
    Relation_Country: string;
    Relation_Phone: string;
    Relation_Email: string;
    SpecialNotes: string;
    BranchID: number | null;
    DepartmentID: number | null;
    CreatedStaffID: number | null;
    CreatedDateTime: string | null;
    UpdatedStaffID: number | null;
    UpdatedDateTime: string | null;
    IsRowDeleted: string;
    InsuranceType: string;
    InsuranceName: string;
    VerifiedBy: string;
}